[Mosaic] Extract Mosaic Wordle related Composable in their own file
